refactor(structure): consolidate features and utils into lib directory
Simplify project structure by reorganizing code following React Router and RTK Query best practices:

**Directory Changes:**
- Move features/search → lib/search.ts (core business logic)
- Move features/recommendations → lib/constants.ts (app configuration)
- Integrate utils/*.js → lib/formatters.ts (unified TypeScript)
- Delete unused search-types.ts (no active references)

**Key Improvements:**
- RTK Query API stays in services/ (official best practice)
- Add formatCompactNumber() for K/M/B number formatting (107.9M)
- Remove duplicate timeFormatter.js (formatDuration already exists)
- Eliminate features/ directory entirely (-75 lines total)

**Files Modified:**
- lib/constants.ts: Add RECOMMENDED_ARTIST_IDS
- lib/formatters.ts: Add formatCompactNumber()
- lib/search.ts: Fuse.js search logic
- pages/search-page.tsx: Update import paths

This consolidation aligns with the constitution's MVP principle by removing unnecessary abstraction layers.
